Rename owners.js module to owners_tree.js
