+++ owners/test/owners_tree.test.js
@@ -16,7 +16,7 @@
 
 const sinon = require('sinon');
 const {LocalRepository} = require('../src/local_repo');
-const {OwnersTree} = require('../src/owners');
+const {OwnersTree} = require('../src/owners_tree');
 const {OwnersParser} = require('../src/parser');
 const {
   OwnersRule,
