+++ owners/test/parser.test.js
@@ -1,6 +1,6 @@
 const sinon = require('sinon');
 const {LocalRepository} = require('../src/local_repo');
-const {OwnersTree} = require('../src/owners');
+const {OwnersTree} = require('../src/owners_tree');
 const {OwnersParser} = require('../src/parser');
 const {
   OwnersRule,
